Move 'guard' namespace to 'lxc'
